feat: misc changes after Anthony review
