+++ great_expectations/rule_based_profiler/config/base.py
@@ -303,6 +303,30 @@ class RuleBasedProfilerConfig(BaseYamlConfig):
     def get_schema_class(cls) -> Type["RuleBasedProfilerConfigSchema"]:  # noqa: F821
         return RuleBasedProfilerConfigSchema
 
+    # noinspection PyUnusedLocal,PyUnresolvedReferences
+    @staticmethod
+    def resolve_config_using_acceptable_arguments(
+        profiler: "RuleBasedProfiler",  # noqa: F821
+        variables: Optional[Dict[str, Any]] = None,
+        rules: Optional[Dict[str, Dict[str, Any]]] = None,
+    ) -> dict:
+        runtime_config: dict = profiler.config.to_dict()
+
+        # If applicable, override config attributes with runtime args
+        if variables:
+            runtime_config["variables"] = variables
+        if rules:
+            runtime_config["rules"] = rules
+
+        runtime_config["variable_count"] = len(runtime_config["variables"])
+        runtime_config["rule_count"] = len(runtime_config["rules"])
+
+        for attr in ("class_name", "module_name", "variables"):
+            runtime_config.pop(attr)
+            logger.debug("Removed unnecessary attr %s from profiler config", attr)
+
+        return runtime_config
+
 
 class RuleBasedProfilerConfigSchema(Schema):
     """
